requirements.txt
asgiref==3.5.2
Django==3.1.6
mysql-connector-python==8.0.31
protobuf==3.20.1
pytz==2022.6
sqlparse==0.4.3
